requirements.txt
python-binance==1.0.19
typer[all]==0.9.0
python-dotenv==1.0.0
click==8.1.7
